pypi metadata: be more precise about python versions
Note: I used the same python versions as for borgbackup because I run
extensive release testing for these versions (which indirectly tests
llfuse also) plus python 3.9 which should also work after fixing llfuse's
cythonized code for it.

There might be a need to drop 3.4 and 3.5 later again because they
are EOL, but for right now, they're still working and tested.